Update README.md
The main objective of this PIA is to calculate the Gestational Age and Estimate Due Date (EDD) for any pregnancy, the gestational age would give the weeks of the pregnancy and the EDD would give the date of the expected deliver day. 
The program would ask the last menstrual period (LMP) date and the current date, in case that the user don’t have the specific date, the program would ask a second question in which the fundal height would be ask (uterus height).